refactor(test): split test_job_executor.py into two files
Split the 717-line test file into:
- test_job_executor.py (408 lines): core functionality tests
  - CommandType to ExecutionRole mapping (7 tests)
  - Initialization (1 test)
  - Context building (1 test)
  - execute method (5 tests)
  - Result mapping (4 tests)

- test_job_executor_dispatch.py (334 lines): dispatch logic tests
  - Equivalence regression (1 test)
  - Governance dispatch (3 tests)
  - Sync dispatch (2 tests)

All 25 tests pass. Both files now comply with 400-line guideline.
